Internal salmonMSE functions for converting operating model inputs and outputs

SOM2MOM() converts a salmon operating model (SOM) to a multi-stock operating model (MOM)

MMSE2SMSE() converts the openMSE output, along with additional state variables recorded in salmonMSE_env, into a salmon MSE object (SMSE) salmonMSE() is the wrapper function that coordinates the simulation and the output.

Usage

MMSE2SMSE(MMSE, SOM, Harvest_MMP, N, Ford)

SOM2MOM(SOM, start = list())

Arguments

MMSE

Object of class MMSE returned from MSEtool

SOM

An object of class SOM

Harvest_MMP

Optional harvest function created by make_Harvest_MMP()

N

Data frame of state variables saved in the salmonMSE_env environment during the simulation

Ford

Data frame of phenotypic trait values saved in the salmonMSE_env environment during the simulation

start

An optional named list to specify the natural smolt production ('Natural') and smolt releases ('Hatchery') at the start of the simulation

Value

MMSE2SMSE: SMSE object

SOM2MOM: MOM object

Examples

if (FALSE) {
# One thousand natural and hatchery smolts in the first year
SOM <- SOM2MOM(SOM, start = list(Natural = 1e3, Hatchery = 1e3))
}