Class "Bio"

The component of the operating model that controls biological dynamics, i.e., natural production.

Details

Various parameters can be stochastic (length nsim) or input as a single numeric (value identical across all simulations).

Slots

Name

Character. Identifying name

nsim

Integer. Number of simulations.

maxage

Integer. The maximum age of the population age structure.

p_mature

Either vector by age (length maxage) or an array with dimension [nsim, maxage, nyears+proyears]. The proportion mature by age.

SRrel

Character, stock-recruit relationship for density-dependent smolt production. Either "BH" (Beverton-Holt) or "Ricker"

capacity_smolt

Vector length nsim. The asymptote of the Beverton-Holt stock-recruit function, or the Ricker maximum for density-dependent natural smolt production from fry life stage. Units of smolts.

kappa

Vector length nsim. The compensation ratio for the stock-recruit function, a.k.a. adult productivity. Natural per-capita production of recruits as the population approaches zero (density-independent component). Units of recruits per egg. In stage-based models, equivalent to the product of smolt productivity (smolts per spawner) and marine survival.

Smax

Vector length nsim. The spawning output that maximizes smolt production in the Ricker stock-recruit function. Units of egg production.

phi

Optional parameter, vector length nsim. Unfished egg production rate per smolt. The alpha parameter of the stock-recruit function will be the ratio of kappa and phi. In stage-based models, the product of marine survival, fecundity, and proportion female. If not provided, phi will be calculated from Mocean_NOS corresponding to the first year.

Mocean_NOS

Either vector by age (length maxage) or an array with dimension [nsim, maxage, nyears+proyears]. Natural mortality of immature natural origin fish. To replicate the SAR parameter of a stage-specific model, set Mocean_NOS[a] = -log(SAR) where a is the age class prior to maturation (and zero for all other ages).

fec

Vector by age (length maxage). Female fecundity of natural origin spawners.

p_female

Numeric. The proportion of females in the spawning population.

Creating Object

Objects can be created by calls of the form new("Bio")

Examples

showClass("Bio")
#> Class "Bio" [package "salmonMSE"]
#> 
#> Slots:
#>                                                                   
#> Name:            Name           nsim         maxage       p_mature
#> Class:      character        numeric        numeric      num.array
#>                                                                   
#> Name:           SRrel capacity_smolt          kappa           Smax
#> Class:      character        numeric        numeric        numeric
#>                                                                   
#> Name:             phi     Mocean_NOS            fec       p_female
#> Class:        numeric      num.array        numeric        numeric
#> 
#> Known Subclasses: "SOM"