Class "Habitat"

The component of the operating model that controls habitat management.

Slots

Name

Character. Identifying name

capacity_smolt_improve

Numeric. The proportional change in the smolt capacity parameter in the projection period. For example, 1.1 represents a 10 percent improvement. For the Beverton-Holt function, this increases the asymptote by 10 percent. For the Ricker function, this increases the maximum smolt production by 10 percent.

kappa_improve

Numeric. The proportional change in the compensation ratio, i.e., changing the smolt productivity, in the projection period. For example, 1.1 represents a 10 percent improvement.

Creating Object

Objects can be created by calls of the form new("Habitat")

Examples

showClass("Habitat")
#> Class "Habitat" [package "salmonMSE"]
#> 
#> Slots:
#>                                                                            
#> Name:                    Name capacity_smolt_improve          kappa_improve
#> Class:              character                numeric                numeric
#> 
#> Known Subclasses: "SOM"