Class "Hatchery"

The component of the operating model that controls the hatchery management.

Details

Various parameters can be stochastic (length nsim) or input as a single numeric (value identical across all simulations).

A description of the fitness parameters and hatchery dynamics is available in the online documentation.

Slots

Name

Character. Identifying name

n_yearling

Numeric. The number of yearlings released. No hatchery is modeled if both n_yearling and n_subyearling are zero.

n_subyearling

Numeric. The number of subyearlings released. No hatchery is modeled if both n_yearling and n_subyearling are zero.

s_prespawn

Numeric. The survival of broodtake prior to egg production. 1 - s_prespawn is the proportion of fish not used for hatchery purposes, e.g., mortality or other resesarch purposes. Used to back-calculate the broodtake from n_yearling and n_subyearling.

s_egg_smolt

Numeric. The survival of eggs to the smolt life stage (for yearling release). Used to back-calculate the broodtake from n_yearling and n_subyearling.

s_egg_subyearling

Numeric. The survival of eggs to subyearling life stage (for subyearling release). Used to back-calculate the broodtake from n_yearling and n_subyearling.

Mocean_HOS

Either vector by age (length maxage) or an array with dimension [nsim, maxage, nyears+proyears]. Natural mortality of immature hatchery origin fish. To replicate the SAR parameter of a stage-specific model, set Mocean_HOS[a] = -log(SAR) for the age class prior to maturation (and zero for all other ages).

gamma

Numeric. The relative fecundity of hatchery origin spawners relative to natural origin spawners.

pmax_NOB

Numeric. The maximum proportion of the natural origin escapement to be used as broodtake.

ptarget_NOB

Numeric. The target proportion of the natural origin broodtake relative to the overall broodtake.

phatchery

Numeric. The proportion of the hatchery origin escapement that return to the hatchery, for example, by removal from spawning grounds or swim-in facilities. These fish are available for broodtake.

premove_HOS

Numeric. The proportion of the hatchery origin escapement to be removed from the spawning grounds (in order to ensure a high proportion of NOS). These fish are not available for broodtake. For example, a value less than one can represent imperfect implementation of weir removal.

fec_brood

Vector of length maxage. The fecundity schedule of broodtake to calculate the total egg production for the hatchery.

fitness_type

Character. Fitness function for hatchery origin spawners. Either "Ford" or "none".

theta

Vector length 2. The optimum phenotype value for the natural and hatchery environments.

rel_loss

Vector length 3. The loss in fitness apportioned among the egg, fry, and smolt life stages. The three values should sum to 1.

pbar_start

Vector length 2. The mean phenotype value in the the natural and hatchery population at the start of the projection.

fitness_variance

Numeric. The variance of the phenotype in the population. Assumed identical between the natural and hatchery environments.

selection_strength

Numeric. The ratio between the fitness standard deviation and the phenotype standard deviation.

heritability

Numeric. The heritability of the phenotypic trait. Between 0-1.

fitness_floor

Numeric. The minimum fitness value in the natural environment.

Creating Object

Objects can be created by calls of the form new("Hatchery")

Examples

showClass("Hatchery")
#> Class "Hatchery" [package "salmonMSE"]
#> 
#> Slots:
#>                                                                
#> Name:                Name         n_yearling      n_subyearling
#> Class:          character            numeric            numeric
#>                                                                
#> Name:          s_prespawn        s_egg_smolt  s_egg_subyearling
#> Class:            numeric            numeric            numeric
#>                                                                
#> Name:          Mocean_HOS              gamma           pmax_NOB
#> Class:          num.array            numeric            numeric
#>                                                                
#> Name:         ptarget_NOB          phatchery        premove_HOS
#> Class:            numeric            numeric            numeric
#>                                                                
#> Name:           fec_brood       fitness_type              theta
#> Class:            numeric          character            numeric
#>                                                                
#> Name:            rel_loss         pbar_start   fitness_variance
#> Class:            numeric            numeric            numeric
#>                                                                
#> Name:  selection_strength       heritability      fitness_floor
#> Class:            numeric            numeric            numeric
#> 
#> Known Subclasses: "SOM"