Class "SOM"

An object containing all the parameters for a salmon operating model (SOM).

Slots

Name

Character. Identifying name

nyears

Integer. The number of historical years

proyears

Integer. The number of projected years

seed

Integer. A random seed to ensure users can reproduce results exactly

nsim

Integer. Number of simulations.

maxage

Integer. The maximum age of the population age structure.

p_mature

Vector of length maxage. The proportion mature by age.

capacity_smolt

Vector length nsim. The asymptote of the Beverton-Holt stock-recruit function for density-dependent natural smolt production from fry life stage.

prod_smolt

Vector length nsim. The productivity parameter of the Beverton-Holt stock-recruit function. Natural per-capita smolt production as the population approaches zero (density-independent component).

SAR

Vector length nsim. Survival to adult return (density-independent, time-invariant).

fec

Numeric. Female fecundity of natural origin spawners.

p_female

Numeric. The proportion of females in the spawning population.

capacity_smolt_improve

Numeric. The proportional change in the smolt capacity parameter in the projection period. For example, 1.1 represents a 10 percent improvement.

prod_smolt_improve

Numeric. The proportional change in the smolt productivity parameter in the projection period. For example, 1.1 represents a 10 percent improvement.

n_yearling

Numeric. The number of yearlings released. No hatchery is modeled if both n_yearling and n_subyearling are zero.

n_subyearling

Numeric. The number of subyearlings released. No hatchery is modeled if both n_yearling and n_subyearling are zero.

s_prespawn

Numeric. The survival of broodtake prior to egg production. 1 - s_prespawn is the proportion of fish not used for hatchery purposes, e.g., mortality or other resesarch purposes. Used to back-calculate the broodtake from n_yearling and n_subyearling.

s_egg_smolt

Numeric. The survival of eggs to the smolt life stage (for yearling release). Used to back-calculate the broodtake from n_yearling and n_subyearling.

s_egg_subyearling

Numeric. The survival of eggs to subyearling life stage (for subyearling release). Used to back-calculate the broodtake from n_yearling and n_subyearling.

gamma

Numeric. The relative fecundity of hatchery origin spawners relative to natural origin spawners.

pmax_NOB

Numeric. The maximum proportion of the natural origin escapement to be used as broodtake.

ptarget_NOB

Numeric. The target proportion of the natural origin broodtake relative to the overall broodtake.

premove_HOS

Numeric. The proportion of the hatchery origin spawners to be removed from the spawning grounds (in order to ensure a high proportion of NOS). This removal is also used to calculate the available hatchery origin fish for broodtake.

fec_brood

Numeric. The fecundity of broodtake.

fitness_type

Character. Fitness function for hatchery origin spawners. Either "Ford" or "none".

theta

Vector length 3.

rel_loss

Vector length 3.

pbar_start

Vector length 2.

fitness_variance

Numeric.

selection_strength

Numeric.

heritability

Numeric.

fitness_floor

Numeric.

u_preterminal

Numeric. The exploitation rate of the immature stock in the pre-terminal fishery.

u_terminal

Numeric. The exploitation rate of the return in the terminal fishery.

m

Numeric. The mark rate of hatchery origin fish, which affects fishery selectivity.

Objects from the Class

Objects can be created by calls of the form new("SOM", Bio, Habitat, Hatchery, Harvest).