Class "SOM"

An object containing all the parameters for a salmon operating model (SOM).

Details

A description of the fitness parameters and hatchery dynamics is available in the online documentation.

Slots

Name

Character. Identifying name

nyears

Integer. The number of historical years

proyears

Integer. The number of projected years

seed

Integer. A random seed to ensure users can reproduce results exactly

nsim

Integer. Number of simulations.

maxage

Integer. The maximum age of the population age structure.

p_mature

Either vector by age (length maxage) or an array with dimension [nsim, maxage, nyears+proyears]. The proportion mature by age.

SRrel

Character, stock-recruit relationship for density-dependent smolt production. Either "BH" (Beverton-Holt) or "Ricker"

capacity_smolt

Vector length nsim. The asymptote of the Beverton-Holt stock-recruit function, or the Ricker maximum for density-dependent natural smolt production from fry life stage. Units of smolts.

kappa

Vector length nsim. The compensation ratio for the stock-recruit function, a.k.a. adult productivity. Natural per-capita production of recruits as the population approaches zero (density-independent component). Units of recruits per egg. In stage-based models, equivalent to the product of smolt productivity (smolts per spawner) and marine survival.

Smax

Vector length nsim. The spawning output that maximizes smolt production in the Ricker stock-recruit function. Units of egg production.

phi

Optional parameter, vector length nsim. Unfished egg production rate per smolt. The alpha parameter of the stock-recruit function will be the ratio of kappa and phi. In stage-based models, the product of marine survival, fecundity, and proportion female. If not provided, phi will be calculated from Mocean_NOS corresponding to the first year.

Mocean_NOS

Either vector by age (length maxage) or an array with dimension [nsim, maxage, nyears+proyears]. Natural mortality of immature natural origin fish. To replicate the SAR parameter of a stage-specific model, set Mocean_NOS[a] = -log(SAR) where a is the age class prior to maturation (and zero for all other ages).

fec

Vector by age (length maxage). Female fecundity of natural origin spawners.

p_female

Numeric. The proportion of females in the spawning population.

capacity_smolt_improve

Numeric. The proportional change in the smolt capacity parameter in the projection period. For example, 1.1 represents a 10 percent improvement. For the Beverton-Holt function, this increases the asymptote by 10 percent. For the Ricker function, this increases the maximum smolt production by 10 percent.

kappa_improve

Numeric. The proportional change in the compensation ratio, i.e., changing the smolt productivity, in the projection period. For example, 1.1 represents a 10 percent improvement.

n_yearling

Numeric. The number of yearlings released. No hatchery is modeled if both n_yearling and n_subyearling are zero.

n_subyearling

Numeric. The number of subyearlings released. No hatchery is modeled if both n_yearling and n_subyearling are zero.

s_prespawn

Numeric. The survival of broodtake prior to egg production. 1 - s_prespawn is the proportion of fish not used for hatchery purposes, e.g., mortality or other resesarch purposes. Used to back-calculate the broodtake from n_yearling and n_subyearling.

s_egg_smolt

Numeric. The survival of eggs to the smolt life stage (for yearling release). Used to back-calculate the broodtake from n_yearling and n_subyearling.

s_egg_subyearling

Numeric. The survival of eggs to subyearling life stage (for subyearling release). Used to back-calculate the broodtake from n_yearling and n_subyearling.

Mocean_HOS

Either vector by age (length maxage) or an array with dimension [nsim, maxage, nyears+proyears]. Natural mortality of immature hatchery origin fish. To replicate the SAR parameter of a stage-specific model, set Mocean_HOS[a] = -log(SAR) for the age class prior to maturation (and zero for all other ages).

gamma

Numeric. The relative reproductive success of hatchery origin spawners (relative to natural origin spawners).

pmax_NOB

Numeric. The maximum proportion of the natural origin escapement to be used as broodtake.

ptarget_NOB

Numeric. The target proportion of the natural origin broodtake relative to the overall broodtake.

phatchery

Numeric. The proportion of the hatchery origin escapement that return to the hatchery, for example, by removal from spawning grounds or swim-in facilities. These fish are available for broodtake.

premove_HOS

Numeric. The proportion of the hatchery origin escapement to be removed from the spawning grounds (in order to ensure a high proportion of NOS). These fish are not available for broodtake. For example, a value less than one can represent imperfect implementation of weir removal.

fec_brood

Vector of length maxage. The fecundity schedule of broodtake to calculate the total egg production for the hatchery.

fitness_type

Character vector length 2. The fitness function to apply in the natural and hatchery environment, respectively. For each, either "Ford" or "none".

theta

Vector length 2. The optimum phenotype value for the natural and hatchery environments.

rel_loss

Vector length 3. The loss in fitness apportioned among the egg, fry, and smolt life stages. The three values should sum to 1.

zbar_start

Vector length 2. The mean phenotype value in the natural and hatchery populations at the start of the projection.

fitness_variance

Numeric. The variance of the phenotypic trait. Assumed identical between the natural and hatchery environments.

selection_strength

Numeric. The ratio between the fitness standard deviation and the phenotype standard deviation.

heritability

Numeric. The heritability of the phenotypic trait. Between 0-1.

fitness_floor

Numeric. The minimum fitness value in the natural and hatchery environments.

u_preterminal

Numeric. The exploitation rate of the immature stock in the pre-terminal fishery.

u_terminal

Numeric. The exploitation rate of the return in the terminal fishery.

m

Numeric. The mark rate of hatchery origin fish, which affects fishery selectivity. Mark-selective fishing (MSF) is implemented if m > 0.

release_mort

Vector length 2. The proportion of released fish that die after release, in the pre-terminal and terminal fishery. Implemented to model mark-selective fishing. Not used if m = 0.

vulPT

Vector length maxage. Vulnerability schedule (between 0-1) in the preterminal fishery.

vulT

Vector length maxage. Vulnerability schedule (between 0-1) in the terminal fishery.

HistSpawner

Array by [nsim, maxage, nyears, 2]. Spawners at age in the natural environment. For the fourth dimension, the first array corresponds to natural origin spawners and the second array corresponds to hatchery origin spawners. Optional

HistN

Array by [nsim, maxage, nyears, 2]. The abundance of immature fish at the beginning of the annual time step. For the fourth dimension, the first array corresponds to natural origin fish and the second array corresponds to hatchery origin fish. Default assumes 1000 smolts (age-1) fish annually.

HistFPT

Vector by historical years (nyears) or an array by dimension [nsim, nyears, 2]. The instantaneous fishing mortality in the preterminal fishery. The first array slice corresponds to F for natural origin fish and the second array slice corresponds to hatchery origin fish. Default is zero.

HistFT

Vector by historical years (nyears) or an array by dimension [nsim, nyears, 2]. The instantaneous fishing mortality in the terminal fishery. The first array slice corresponds to F for natural origin fish and the second array slice corresponds to hatchery origin fish. Default is zero.

Objects from the Class

Objects can be created by calls of the form new("SOM", Bio, Habitat, Hatchery, Harvest, Historical).