Plot life history groups

Plot the annual proportions of life history groups at various life stages

Usage

plot_LHG(
  SMSE,
  var = "NOS",
  s = 1,
  FUN = median,
  figure = TRUE,
  xlab = "Projection Year",
  ylab = paste("Proportion", var),
  LHG_names
)

Arguments

SMSE

Class SMSE object returned by salmonMSE()

var

Character. Slot for the state variables in SMSE@Misc$LHG[[1]].

s

Integer. Population index for multi-population model (e.g., s = 1 is the first population in the model)

FUN

Summarizing function across simulations, typically median() or mean()

figure

Logical, whether to generate a figure (set to FALSE if only using the function to return the data matrix)

xlab

Character. Name of time variable for the figure

ylab

Character. Name of the state variable for the figure

LHG_names

Character. Vector of names for the life history groups

See also