Plot output from salmonMSE

Various functions that plot the state variables from salmonMSE projections:

plot_statevar_ts() produces a time series for all simulations, or with confidence intervals

plot_statevar_hist() produces a histogram across all simulations for a particular year

Usage

plot_statevar_ts(
  SMSE,
  var = "PNI",
  s = 1,
  xlab = "Projection Year",
  quant = FALSE,
  ylab = var,
  ylim
)

plot_statevar_hist(SMSE, var = "PNI", s = 1, y, xlab = var, ...)

Arguments

SMSE

Class SMSE object returned by salmonMSE()

var

Character. Slot for the state variable in SMSE object. See slotNames(SMSE) for options.

s

Integer. Population index for multi-population model (e.g., s = 1 is the first population in the model)

xlab

Character. Name of time variable for the figure

quant

Logical, whether to plot individual simulations (FALSE) or the median with 95 percent confidence intervals (TRUE)

ylab

Character. Name of the state variable for the figure

ylim

Vector. Y-axis limits

y

Integer. Projection year for the state variable to plot the histogram

Value

Base graphics