Run salmonMSE

salmonMSE() runs a salmon management strategy evaluation through the following steps:

Converts a salmon operating model (SOM) to a multi-stock operating model (MOM) via SOM2MOM()

Creates a harvest management procedure specifying the harvest control rule

Generates the historical reconstruction of the state variables

Runs projection (if Hist = FALSE)

Converts the openMSE output, along with additional state variables recorded in salmonMSE_env, into a salmon MSE object (SMSE) via MMSE2SMSE()

Usage

salmonMSE(SOM, Hist = FALSE, silent = FALSE, trace = FALSE, convert = TRUE)

Arguments

SOM

An object of class SOM

Hist

Logical, whether to stop the function stop after historical simulations? Returns a list containing all historical data

silent

Logical, whether to report progress in console bar

trace

Logical, whether to report additional messages from openMSE

convert

Logical, whether to convert the output into a salmon MSE (SMSE) object

Value

If Hist = TRUE a multiHist object (list). Otherwise, if convert = TRUE, a SMSE object or if convert = FALSE, a MMSE object.