Internal salmonMSE functions for converting operating model inputs and outputs

SOM2MOM() converts a salmon operating model (SOM) to a multi-stock operating model (MOM)

make_Stock() creates the Stock object (openMSE) corresponding to salmon life stage

make_Fleet() creates the Fleet object (openMSE) corresponding to the fishery that interacts with the various salmon life stages

multiHist2SHist() converts the openMSE historical reconstruction into a salmon Hist object (SHist)

MMSE2SMSE() converts the openMSE projection output, along with additional state variables recorded in salmonMSE_env, into a salmon MSE object (SMSE)

salmonMSE() is the wrapper function that coordinates the simulation and the output.

Usage

MMSE2SMSE(MMSE, SOM, Harvest_MMP, N, Ford, state)

SOM2MOM(SOM, check = TRUE)

make_Stock(SOM, NOS = TRUE, stage = c("immature", "return", "escapement"))

make_Fleet(SOM, NOS = TRUE, stage = c("immature", "return", "escapement"))

multiHist2SHist(multiHist, SOM, check = TRUE)

Arguments

MMSE

Object of class MMSE returned from MSEtool

SOM

An object of class SOM

Harvest_MMP

Optional harvest function created by make_Harvest_MMP()

N

Data frame of abundance at age saved in the salmonMSE_env environment during the simulation

Ford

Data frame of phenotypic trait values saved in the salmonMSE_env environment during the simulation

state

Data frame of state variables saved in the salmonMSE_env environment during the simulation

check

Logical, whether to check the SOM object using check_SOM()

NOS

Logical, whether the Stock or Fleet object corresponds to natural origin or hatchery origin fish

stage

Character indicating the corresponding salmon life stage of the Stock or Fleet object

multiHist

Class multiHist object returned from MSEtool

Value

MMSE2SMSE: SMSE object

SOM2MOM: MOM object

make_Stock: List containing a Stock object and accompanying custom parameters list

make_Stock: List containing a Fleet object and accompanying custom parameters list

multiHist2SHist: SHist object